4.Fast Api/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
alembic
PyMySQL
python-dotenv
pydantic
pytest
